Provide a libctl3 symlink for the mpb package
Per alexp, the mpb package is looking in /usr/share/libctl3, but
the "libctl3" package installs things in /usr/share/libctl.  This
creates the necessary symlink. (Trac: #1451)

Add a Conflicts: for mpb >= 5.1~rc1-1 which contains the fix that
renders the symlink unnecessary.  Once that version is available,
we can drop the Conflicts: and the .links file.